+++ src/lib.rs
@@ -1,15 +1,14 @@
 pub mod diagnostic;
 pub mod prelude;
+pub mod schedule;
 
 mod options;
 mod plugin;
 mod system;
-mod system_set;
 mod wrapper;
 
 pub use options::*;
 pub use plugin::*;
-pub use system_set::*;
 pub use wrapper::*;
 
 pub use pixels;