+++ src/schedule.rs
@@ -1,0 +1,7 @@
+use bevy::ecs::schedule::ScheduleLabel;
+
+#[derive(ScheduleLabel, Clone, Debug, PartialEq, Eq, Hash)]
+pub struct Draw;
+
+#[derive(ScheduleLabel, Clone, Debug, PartialEq, Eq, Hash)]
+pub struct Render;